working request cannot delete

+++ routers/admin.js
@@ -77,10 +77,18 @@ router.get('/c_admin_request_manage', wrapper.asyncMiddleware(async (req, res, n
 router.post('/c_admin_request_delete', wrapper.asyncMiddleware(async (req, res, next) =>{
 	var rid_delete = req.body.rid;
 	var pid_delete = req.body.pid;
+	var State = await db.getQueryResult(`SELECT State FROM Request WHERE RID = '${RID}' and PID = '${PID}'`);
 	// console.log(""+rid_delete+" "+pid_delete);
-	var sql = `DELETE FROM Request WHERE PID = '${pid_delete}' and RID = '${rid_delete}'`;
-	const result = await db.getQueryResult(sql);
-	res.json(result);
+	if(State[0].State == 0 || State[0].State == 4){
+		var sql = `DELETE FROM Request WHERE PID = '${pid_delete}' and RID = '${rid_delete}'`;
+		const result = await db.getQueryResult(sql);
+		res.json(result);
+	}else{
+		res.json(400, {
+                   error: 1,
+                   msg: "request running"
+		});
+	}
 }));
 router.post('/c_admin_request_update', wrapper.asyncMiddleware(async (req, res, next) =>{
 	var RID = req.body.rid;
@@ -122,6 +130,7 @@ router.post('/c_admin_team_delete', wrapper.asyncMiddleware(async (req, res, nex
 	var TeamName = req.body.teamname;
 	var check_sql = `SELECT RID FROM Attend WHERE TName = '${TeamName}'`
 	var check = await db.getQueryResult(check_sql);
+
 	if(check.length >= 1){
 		var sql = `DELETE FROM TeamList WHERE TeamName = '${TeamName}'`;
 		const result = await db.getQueryResult(sql);
